+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to the "linecount" extension will be documented in this file.
 
+## Version 0.1.7 (2017-06-03)
+
+1. Add output CSV,markdown file type.
+
+2. Add output files sort.
+
+
 ## Version 0.1.6 (2017-03-01)
 
 1. To determine whether the binary file.